Changes in model selection choice.
Choice between exponential and weibull distribution describing the various hazard parameters for the carepath model was written as a separate function for each parameter.

Rewritten as:
selectModelOnDate(paraList,thisDate,paraName)
where paraList is the list of fitted survival models for each day of the incoming data, thisDate is the day we are looking at, and paraName is the parameter of interest.

e.g. selectModelOnDate(UKFparaList,as.Date("2020-05-01"),"disICU") gives the model describing the discharge from the ICU, for data until May 1 as measured in the UKF.

The deprecated functions (e.g. getDeltaIDistr) are still there, but should be removed from the rest of the code in the near future.