Adding admission rate estimator
admRateEstimator estimate the proportion of confirmed cases that are admitted to the hospital in question, based on the length of stay distribution produced by the care path model, in combination with the observed number of occupied beds and number of reported cases.

To get the LOS distribution, it runs getLOSdistr(), which simulates 10.000 patients (default, can be changed), all admitted on the same day. Reversing this distribution multiplied by the incidence in the last part of the epicurve gives the total number of confirmed cases underlying these hospitalised patients. The rate is then easily calculated.

I added getTimeLineFromCE() to convert the change events (CE) into the number of occupied bed on each day.
It basically does the same as parts of extractRunInfo(), probably more efficiently. I still need to merge the new function into there, potentially deprecating the old function.

Added the dots for the observed bed occupancy for ICU and General ward combined (black) to the getBedShadedPlot()